Modyfikacja rest kontrolerów

+++ src/main/java/pl/medm/javadev/service/LectureService.java
@@ -6,10 +6,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import pl.medm.javadev.model.Lecture;
+import pl.medm.javadev.model.User;
 import pl.medm.javadev.repository.LectureRepository;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class LectureService {
@@ -21,11 +23,11 @@ public class LectureService {
         this.lectureRepository = lectureRepository;
     }
 
-    public List<Lecture> findAll() {
+    public List<Lecture> getAllLectures() {
         return lectureRepository.findAll();
     }
 
-    public ResponseEntity<?> saveLecture(Lecture lecture) {
+    public ResponseEntity<?> createLecture(Lecture lecture) {
         if (lectureRepository.existsByTitle(lecture.getTitle())) {
             return ResponseEntity.status(HttpStatus.CONFLICT).build();
         } else {
@@ -39,18 +41,61 @@ public class LectureService {
         }
     }
 
-    public ResponseEntity<?> getLecture(Long id) {
+    public ResponseEntity<?> findLectureById(Long id) {
         return lectureRepository.findById(id)
                 .map(ResponseEntity::ok)
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    public ResponseEntity<?> deleteUser(Long id) {
-        if(lectureRepository.existsById(id)) {
-            lectureRepository.deleteById(id);
-            return ResponseEntity.noContent().build();
+    public ResponseEntity<Lecture> updateLectureById(Long id, Lecture newLecture) {
+        if (lectureRepository.existsById(id)) {
+            Lecture lecture = lectureRepository.findById(id).get();
+            lecture.setTitle(newLecture.getTitle());
+            lecture.setDescription(newLecture.getDescription());
+            lecture.setLecturer(newLecture.getLecturer());
+            lectureRepository.save(lecture);
+            return ResponseEntity.ok(lecture);
         } else {
             return ResponseEntity.notFound().build();
         }
     }
+
+    public ResponseEntity<?> deleteLectureById(Long id) {
+        if (lectureRepository.existsById(id)) {
+            if (lectureRepository.findById(id).get().getCompleted()) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+            lectureRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
+    }
+
+    public ResponseEntity<List<User>> getAllUserById(Long id) {
+        Optional<Lecture> lecture = lectureRepository.findById(id);
+        if (lecture.isPresent()) {
+            List<User> users = lecture.get().getUsers();
+            return ResponseEntity.ok(users);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
+    public ResponseEntity<User> saveUser(Long id, User user) {
+        if (lectureRepository.existsById(id)) {
+            Lecture lecture = lectureRepository.findById(id).get();
+            lecture.getUsers().add(user);
+            lectureRepository.save(lecture);
+            URI location = ServletUriComponentsBuilder
+                    .fromCurrentRequest()
+                    .path("/{id}")
+                    .buildAndExpand(user.getId())
+                    .toUri();
+            return ResponseEntity.created(location).body(user);
+        } else {
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
+    }
+
+
 }
